requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
docutils==0.18.1
idna==3.3
Kivy @ git+https://github.com/kivy/kivy.git@4d83ec5baa28f956a130398d19230ff1ef895098
Kivy-Garden==0.1.4
kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip
Pillow==9.0.1
Pygments==2.11.2
requests==2.27.1
urllib3==1.26.8
watchdog==2.1.6
